[INDY-2154] move token tracker and batch handler into sovtokenfees
