+++ sovtoken/sovtoken/main.py
@@ -1,12 +1,10 @@
 from sovtoken.request_handlers.batch_req_handler.token_batch_handler import TokenBatchHandler
-from sovtoken.request_handlers.batch_req_handler.tracker_batch_handler import TrackerBatchHandler
 from sovtoken.request_handlers.batch_req_handler.utxo_batch_handler import UTXOBatchHandler
 from sovtoken.request_handlers.read_req_handler.get_utxo_handler import GetUtxoHandler
 from sovtoken.request_handlers.write_request_handler.mint_handler import MintHandler
 from sovtoken.request_handlers.write_request_handler.xfer_handler import XferHandler
 from sovtoken.utxo_cache import UTXOCache
 
-from plenum.common.ledger_uncommitted_tracker import LedgerUncommittedTracker
 from state.pruning_state import PruningState
 
 from storage.helper import initKeyValueStorage
@@ -57,10 +55,6 @@ def init_storages(node):
                                            node.config.utxoCacheStorage,
                                            node.dataLocation,
                                            node.config.utxoCacheDbName)))
-    token_tracker = LedgerUncommittedTracker(token_state.committedHeadHash,
-                                             token_ledger.uncommitted_root_hash,
-                                             token_ledger.size)
-    node.db_manager.register_new_tracker(TOKEN_LEDGER_ID, token_tracker)
 
 
 def init_token_ledger(node):
@@ -98,7 +92,6 @@ def register_req_handlers(node):
 def register_batch_handlers(node):
     node.write_manager.register_batch_handler(UTXOBatchHandler(node.db_manager), add_to_begin=True)
     node.write_manager.register_batch_handler(TokenBatchHandler(node.db_manager), add_to_begin=True)
-    node.write_manager.register_batch_handler(TrackerBatchHandler(node.db_manager))
     node.write_manager.register_batch_handler(node.write_manager.audit_b_handler,
                                               ledger_id=TOKEN_LEDGER_ID)
